pyproject.toml
[project]
name = "openapi-rulesbank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "jsonschema>=4.17.3",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.1.2",
    "openai>=2.28.0",
    "pydantic>=2.12.5",
    "pytest>=7.1.2",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "torch>=2.0.0",
]

[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]  
